requirements.txt
fastapi==0.115.0
uvicorn==0.23.2
validators==0.22.0
urllib3==2.2.3
redis==5.2.1
requests==2.31.0
pytest==7.4.0
pytest-asyncio
pymongo==4.9
PyJWT==2.9.0
pytz==2024.2
cryptography==44.0.0
PyPika==0.48.9
pika==1.3.2
motor==3.6.1
pydantic==2.9.2
pydantic-settings==2.7.1
pydantic[email]
python-dotenv==1.0.1
httpx==0.28.1
python-multipart==0.0.20
pycryptodome==3.21.0
odmantic==1.0.2
pyzipper==0.3.6
pandas
boto3
strawberry-graphql
jinja2
python-jose[cryptography]
slowapi==0.1.9
redis>=5.0.0
